requirements.txt
asgiref==3.5.2
Django==4.0.6
django-mysql==3.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
install==1.3.5
mysqlclient==2.0.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
django-filter==22.1
django-cors-headers==3.13.0
gunicorn==20.1.0